add support for modifing a todo from the all list
This was not working previously because we were trying to use the
currentList to delete items. However, the all list is a virtual list
so this concept doesn't map so well. We can add to the all list, which
means placing the new items in the default list. But deleting is
tricky because the item could be from another list other than the
default.

This was solved by adding a two-way binding between todo lists and
todo items. Now we can find the list from the item, and then call the
relevant function,